+++ src/themes/teams/components/Divider/dividerStyles.ts
@@ -1,10 +1,6 @@
 import { childrenExist, pxToRem } from '../../../../lib'
-import {
-  IComponentPartStylesInput,
-  ICSSInJSStyle,
-  ICSSPseudoElementStyle,
-} from '../../../../../types/theme'
-import { IDividerPropsWithDefaults } from '../../../../components/Divider/Divider'
+import { ComponentSlotStylesInput, ICSSInJSStyle, ICSSPseudoElementStyle } from '../../../types'
+import { DividerPropsWithDefaults } from '../../../../components/Divider/Divider'
 
 const dividerBorderStyle = (size, color): ICSSInJSStyle => ({
   height: `${size + 1}px`,
@@ -20,7 +16,7 @@ const beforeAndAfter = (size, type, variables): ICSSPseudoElementStyle => ({
   }),
 })
 
-const dividerStyles: IComponentPartStylesInput<IDividerPropsWithDefaults, any> = {
+const dividerStyles: ComponentSlotStylesInput<DividerPropsWithDefaults, any> = {
   root: ({ props, variables }): ICSSInJSStyle => {
     const { children, fitted, size, type, important, content } = props
     return {