+++ test/specs/commonTests/isConformant.tsx
@@ -1,30 +1,27 @@
 import * as _ from 'lodash'
 import * as React from 'react'
-import { mount as enzymeMount } from 'enzyme'
+import { ReactWrapper } from 'enzyme'
 import * as ReactDOMServer from 'react-dom/server'
-import { ThemeProvider } from 'react-fela'
 
 import isExportedAtTopLevel from './isExportedAtTopLevel'
-import { assertBodyContains, consoleUtil, syntheticEvent } from 'test/utils'
+import {
+  assertBodyContains,
+  consoleUtil,
+  mountWithProvider as mount,
+  syntheticEvent,
+} from 'test/utils'
 import helpers from './commonHelpers'
 
 import * as stardust from 'src/'
-import { felaRenderer } from 'src/lib'
+
 import { FocusZone } from 'src/lib/accessibility/FocusZone'
 import { FOCUSZONE_WRAP_ATTRIBUTE } from 'src/lib/accessibility/FocusZone/focusUtilities'
 
-export interface IConformant {
+export interface Conformant {
   eventTargets?: object
   requiredProps?: object
   exportedAtTopLevel?: boolean
   rendersPortal?: boolean
-}
-
-export const mount = (node, options?) => {
-  return enzymeMount(
-    <ThemeProvider theme={{ renderer: felaRenderer }}>{node}</ThemeProvider>,
-    options,
-  )
 }
 
 /**
@@ -32,11 +29,11 @@ export const mount = (node, options?) => {
  * @param {React.Component|Function} Component A component that should conform.
  * @param {Object} [options={}]
  * @param {Object} [options.eventTargets={}] Map of events and the child component to target.
- * @param {boolean} [options.exportedAtTopLevel=false] Is this component exported as top level API
+ * @param {boolean} [options.exportedAtTopLevel=false] Is this component exported as top level API?
  * @param {boolean} [options.rendersPortal=false] Does this component render a Portal powered component?
  * @param {Object} [options.requiredProps={}] Props required to render Component without errors or warnings.
  */
-export default (Component, options: IConformant = {}) => {
+export default (Component, options: Conformant = {}) => {
   const {
     eventTargets = {},
     exportedAtTopLevel = true,
@@ -48,7 +45,7 @@ export default (Component, options: IConformant = {}) => {
   const componentType = typeof Component
 
   // This is added because the component is mounted
-  const getComponent = wrapper => {
+  const getComponent = (wrapper: ReactWrapper) => {
     // FelaTheme wrapper and the component itself:
     let component = wrapper
       .childAt(0)