+++ test/specs/components/RadioGroup/RadioGroupItem-test.ts
@@ -1,6 +1,6 @@
 import { isConformant, handlesAccessibility } from 'test/specs/commonTests'
 
-import { RadioGroupItem } from 'src/components/RadioGroup'
+import RadioGroupItem from 'src/components/RadioGroup/RadioGroupItem'
 
 describe('RadioGroupItem', () => {
   describe('isConformant', () => {